% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{clear_sulcimap_cache}
\alias{clear_sulcimap_cache}
\title{Clear sulcimap cached assets}
\usage{
clear_sulcimap_cache()
}
\value{
Logical TRUE if cache removed or did not exist
}
\description{
Clear sulcimap cached assets
}
