% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{.ensure_sulcimap_assets}
\alias{.ensure_sulcimap_assets}
\title{Ensure bundled assets are available locally (unzips on first use)}
\usage{
.ensure_sulcimap_assets()
}
\value{
Path to the directory containing extracted assets
}
\description{
Ensure bundled assets are available locally (unzips on first use)
}
\keyword{internal}
