% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_assets_light}
\alias{get_assets_light}
\title{Lightweight assets for light example
Unzips a tiny assets bundle and returns the directory path.}
\usage{
get_assets_light()
}
\value{
A character scalar: path to the unzipped assets directory.
}
\description{
Lightweight assets for light example
Unzips a tiny assets bundle and returns the directory path.
}
