% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{get_sulcimap_assets_dir}
\alias{get_sulcimap_assets_dir}
\title{Path to sulcimap extracted asset directory}
\usage{
get_sulcimap_assets_dir()
}
\value{
Character scalar: full path to the root of extracted assets
}
\description{
This will unzip the bundled assets into a per-user cache
the first time it is called then return the cache path on subsequent calls
}
