% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{get_sulcus_image}
\alias{get_sulcus_image}
\title{Full path to a bundled PNG (after extraction)}
\usage{
get_sulcus_image(name)
}
\arguments{
\item{name}{Relative path inside the assets (e.g. "leftlat/S.C._left.png")}
}
\value{
Character scalar: full path
}
\description{
Full path to a bundled PNG (after extraction)
}
