% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{list_sulci_images}
\alias{list_sulci_images}
\title{List bundled PNGs (after extraction)}
\usage{
list_sulci_images()
}
\value{
Character vector of relative PNG paths
}
\description{
List bundled PNGs (after extraction)
}
