% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sulcimap-package.R
\docType{package}
\name{sulcimap-package}
\alias{sulcimap}
\alias{sulcimap-package}
\title{sulcimap: Plot sulcal measures on brain}
\description{
Create composite, publication-quality brain figures to map sulcal values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/<your-user>/sulcimap}
  \item Report bugs at \url{https://github.com/<your-user>/sulcimap/issues}
}

}
\author{
\strong{Maintainer}: Mahan Shafie \email{mahan.shafie@unito.it}

}
\keyword{internal}
