% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{get_sulcimap_assets_dir}
\alias{get_sulcimap_assets_dir}
\title{Path to sulcimap extracted asset directory}
\usage{
get_sulcimap_assets_dir()
}
\value{
Character scalar: full path to the root of extracted assets
}
\description{
Unzips the bundled assets into a cache (persistent for users, temporary under checks)
}
