% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sulcimap-package.R
\docType{package}
\name{sulcimap-package}
\alias{sulcimap}
\alias{sulcimap-package}
\title{sulcimap: Mapping Brain Folding Patterns}
\description{
Create composite, publication-quality brain figures to map sulcal values.
}
\author{
\strong{Maintainer}: Mahan Shafie \email{mahan.shafie@unito.it}

}
\keyword{internal}
